Added changes for a smooth reading experience.
There's less clutter if get rid of quotation marks around book titles, it's better to italicize it. Also corrected thread because it wasn't a thread, it was a tweet. In the section Why Attack Haskell, the "you may be wondering: Why am I attacking Haskell" causes confusion to the reader because they're not the one attacking Haskell, it's you. So the question must be addressed to you rather than to themselves.  I removed personally because I is enough. And I also restructured the sentence: "Too many space leaks have been caused by this function" to make it stronger.